# admission_module/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Admission and Enrollment API",
    "description": "Backend Module for the College ERP System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/brochures/request": {
      "post": {
        "tags": [
          "Brochures"
        ],
        "summary": "Request Brochure",
        "description": "Step 1: An applicant requests a brochure for a specific course.",
        "operationId": "request_brochure_api_v1_brochures_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrochureRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrochureRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/brochures/pay": {
      "post": {
        "tags": [
          "Brochures"
        ],
        "summary": "Pay For Brochure",
        "description": "Step 2: Applicant pays the required 200rs fee for the brochure.\nThis also updates the Brochure Status to 'Payment Done'.",
        "operationId": "pay_for_brochure_api_v1_brochures_pay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroPaymentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroPaymentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/submit": {
      "post": {
        "tags": [
          "Applications"
        ],
        "summary": "Submit Application",
        "description": "Step 1 (Application): Applicant submits their entire application payload.\nStatus starts as 'draft' until documents are uploaded and finalized.",
        "operationId": "submit_application_api_v1_applications_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{application_id}/finalize": {
      "post": {
        "tags": [
          "Applications"
        ],
        "summary": "Finalize Application Final",
        "description": "Step 3 (Application): Applicant finalizes the application.\nChecks for mandatory documents before moving from 'draft' to 'submitted'.",
        "operationId": "finalize_application_final_api_v1_applications__application_id__finalize_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{application_id}/status": {
      "get": {
        "tags": [
          "Applications"
        ],
        "summary": "Get Application Status Tracking",
        "description": "Applicant can track the progress of their application through status logs.",
        "operationId": "get_application_status_tracking_api_v1_applications__application_id__status_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationStatusLogOut"
                  },
                  "title": "Response Get Application Status Tracking Api V1 Applications  Application Id  Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/applications/{application_id}": {
      "get": {
        "tags": [
          "Admin Review"
        ],
        "summary": "Get Application Details",
        "description": "Get detailed information about a specific application.",
        "operationId": "get_application_details_api_v1_admin_applications__application_id__get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/review": {
      "post": {
        "tags": [
          "Admin Review"
        ],
        "summary": "Submit Application Review",
        "description": "Step 5: Admin reviews application and approves/rejects/requests revision.",
        "operationId": "submit_application_review_api_v1_admin_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminReviewCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/document/verify": {
      "post": {
        "tags": [
          "Admin Review"
        ],
        "summary": "Verify Application Document",
        "description": "Admin verifies or rejects a specific document uploaded by the applicant.",
        "operationId": "verify_application_document_api_v1_admin_document_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentVerificationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentVerificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminReviewCreate": {
        "properties": {
          "application_id": {
            "type": "integer",
            "title": "Application Id"
          },
          "admin_id": {
            "type": "integer",
            "title": "Admin Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "remarks": {
            "type": "string",
            "title": "Remarks"
          }
        },
        "type": "object",
        "required": [
          "application_id",
          "admin_id",
          "action",
          "remarks"
        ],
        "title": "AdminReviewCreate"
      },
      "AdminReviewOut": {
        "properties": {
          "review_id": {
            "type": "integer",
            "title": "Review Id"
          },
          "application_id": {
            "type": "integer",
            "title": "Application Id"
          },
          "admin_id": {
            "type": "integer",
            "title": "Admin Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "remarks": {
            "type": "string",
            "title": "Remarks"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "review_id",
          "application_id",
          "admin_id",
          "action",
          "remarks",
          "timestamp"
        ],
        "title": "AdminReviewOut"
      },
...
      "ApplicationCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "brochure_id": {
            "type": "integer",
            "title": "Brochure Id"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          },
          "admission_year": {
            "type": "integer",
            "title": "Admission Year"
          },
          "applicant": {
            "$ref": "#/components/schemas/ApplicantDetailsCreate"
          },
          "parent_details": {
            "$ref": "#/components/schemas/ParentDetailsCreate"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationDetailsCreate"
            },
            "type": "array",
            "title": "Education"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "brochure_id",
          "course_name",
          "admission_year",
          "applicant",
          "parent_details",
          "education"
        ],
        "title": "ApplicationCreate"
      },
      "ApplicationOut": {
        "properties": {
          "application_id": {
            "type": "integer",
            "title": "Application Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "brochure_id": {
            "type": "integer",
            "title": "Brochure Id"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          },
          "current_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Status",
            "default": "draft"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "default": true
          },
          "is_active_or_not": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active Or Not",
            "default": true
          },
          "submission_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Date"
          },
          "admission_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admission Year"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "applicant_details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicantDetailsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "parent_details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParentDetailsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "education_details": {
            "items": {
              "$ref": "#/components/schemas/EducationDetailsOut"
            },
            "type": "array",
            "title": "Education Details",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "application_id",
          "user_id",
          "brochure_id",
          "course_name",
          "created_at"
        ],
        "title": "ApplicationOut"
      },
      "ApplicationStatusLogOut": {
        "properties": {
          "status_log_id": {
            "type": "integer",
            "title": "Status Log Id"
          },
          "application_id": {
            "type": "integer",
            "title": "Application Id"
          },
          "changed_by": {
            "type": "integer",
            "title": "Changed By"
          },
          "changed_role": {
            "type": "string",
            "title": "Changed Role"
          },
          "status_id": {
            "type": "string",
            "title": "Status Id"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "changed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Changed At"
          }
        },
        "type": "object",
        "required": [
          "status_log_id",
          "application_id",
          "changed_by",
          "changed_role",
          "status_id",
          "remark",
          "changed_at"
        ],
        "title": "ApplicationStatusLogOut"
      },
...
      "BroPaymentCreate": {
        "properties": {
          "brochure_id": {
            "type": "integer",
            "title": "Brochure Id"
          }
        },
        "type": "object",
        "required": [
          "brochure_id"
        ],
        "title": "BroPaymentCreate"
      },
      "BroPaymentOut": {
        "properties": {
          "payment_id": {
            "type": "integer",
            "title": "Payment Id"
          },
          "brochure_id": {
            "type": "integer",
            "title": "Brochure Id"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "paid_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          }
        },
        "type": "object",
        "required": [
          "payment_id",
          "brochure_id",
          "amount",
          "status"
        ],
        "title": "BroPaymentOut"
      },
      "BrochureRequestCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "course_name"
        ],
        "title": "BrochureRequestCreate"
      },
      "BrochureRequestOut": {
        "properties": {
          "brochure_id": {
            "type": "integer",
            "title": "Brochure Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          },
          "brochure_status": {
            "type": "string",
            "title": "Brochure Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "brochure_id",
          "user_id",
          "course_name",
          "brochure_status",
          "created_at",
          "updated_at"
        ],
        "title": "BrochureRequestOut"
      },
...
      "DocumentVerificationCreate": {
        "properties": {
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "admin_id": {
            "type": "integer",
            "title": "Admin Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "admin_id",
          "status"
        ],
        "title": "DocumentVerificationCreate"
      },
      "DocumentVerificationOut": {
        "properties": {
          "verification_id": {
            "type": "integer",
            "title": "Verification Id"
          },
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "admin_id": {
            "type": "integer",
            "title": "Admin Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "required": [
          "verification_id",
          "document_id",
          "admin_id",
          "status",
          "remarks"
        ],
        "title": "DocumentVerificationOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ParentDetailsCreate": {
        "properties": {
          "parent_relationship": {
            "type": "string",
            "title": "Parent Relationship"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mother_name": {
            "type": "string",
            "title": "Mother Name"
          },
          "occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "mobile_number": {
            "type": "string",
            "title": "Mobile Number"
          },
          "annual_income": {
            "type": "number",
            "title": "Annual Income"
          },
          "guardian_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Name"
          },
          "guardian_occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Occupation"
          },
          "guardian_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Email"
          },
          "guardian_mobile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Mobile"
          }
        },
        "type": "object",
        "required": [
          "parent_relationship",
          "name",
          "mother_name",
          "mobile_number",
          "annual_income"
        ],
        "title": "ParentDetailsCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}